tree with conditions and actions
