+++ crates/world_state/src/behavior/condition.rs
@@ -1,6 +1,35 @@
+use coordinate_systems::Field;
+use linear_algebra::{Vector2, vector};
 use types::primary_state::PrimaryState;
 
 use crate::behavior::node::Blackboard;
+
+pub fn is_close_to_ball(blackboard: &mut Blackboard) -> bool {
+    if let Some(ball) = &blackboard.world_state.ball {
+        ball.ball_in_ground.coords().norm()
+            < blackboard.parameters.kicking.distance_for_kick_hysteresis
+    } else {
+        false
+    }
+}
+
+pub fn is_close_to_goal(blackboard: &mut Blackboard) -> bool {
+    if let Some(ground_to_field) = blackboard.world_state.robot.ground_to_field {
+        let field_to_ground = ground_to_field.inverse();
+
+        let goal_position: Vector2<Field> = vector!(blackboard.field_dimensions.length / 2.0, 0.0);
+
+        let target_position = (field_to_ground * goal_position).as_point();
+
+        target_position.coords().norm()
+            < blackboard
+                .parameters
+                .kicking
+                .goal_distance_kick_power_threshold
+    } else {
+        false
+    }
+}
 
 pub fn is_closest_to_ball(_blackboard: &mut Blackboard) -> bool {
     // TODO
