+++ crates/world_state/src/behavior/mod.rs
@@ -3,3 +3,4 @@ pub mod behavior_tree;
 pub mod condition;
 pub mod node;
 pub mod tree;
+pub mod kick_actions;
