+++ crates/world_state/src/behavior/tree.rs
@@ -1,14 +1,11 @@
-use types::primary_state::PrimaryState;
+use types::{motion_command::KickPower, primary_state::PrimaryState};
 
 use crate::{
     action,
     behavior::{
-        action::{injected_motion_command, kick, leuchtturm, prepare, stand, stand_up},
-        behavior_tree::Node,
-        condition::{
-            has_ball_position, is_closest_to_ball, is_fallen, is_goalkeeper, is_primary_state,
-        },
-        node::Blackboard,
+        action::{injected_motion_command, leuchtturm, prepare, stand, stand_up}, behavior_tree::Node, condition::{
+            has_ball_position, is_close_to_ball, is_close_to_goal, is_closest_to_ball, is_fallen, is_goalkeeper, is_primary_state
+        }, kick_actions::kicking, node::Blackboard
     },
     condition, negation, selection, sequence,
 };
@@ -70,7 +67,17 @@ fn search_subtree() -> Node<Blackboard> {
 }
 
 fn striker_subtree() -> Node<Blackboard> {
-    action!(kick)
+    selection!(
+        sequence!(
+            negation!(condition!(is_close_to_ball)),
+            action!(walk_to_ball)
+        ),
+        sequence!(
+            condition!(is_close_to_goal),
+            action!(kicking, KickPower::Rumpelstilzchen)
+        ),
+        action!(kicking, KickPower::Schlong)
+    )
 }
 
 fn supporter_subtree() -> Node<Blackboard> {
